+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - Removed instances of TryEnter from presenters
+ - Adding OccupancyManager to rooms
+ - Using new OccupancyManager features for occupancy state
 
 ## [15.0.1] - 2021-08-23
 ### Changed